requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
youtube-transcript-api
streamlit
google-generativeai
langchain
pathlib
jupyter
huggingface_hub
